backend/requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
arcade==3.3.3
asgiref==3.8.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.1
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
dash==3.2.0
distro==1.9.0
Django==5.1.6
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docx==0.2.4
et_xmlfile==2.0.0
filetype==1.2.0
Flask==3.1.2
flatbuffers==25.2.10
fonttools==4.58.1
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-genai==1.54.0
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
groq==0.18.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.14.0
httpcore==1.0.7
httplib2==0.31.0
httpx==0.28.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==1.1.2
langchain-core==1.1.2
langchain-google-genai==4.0.0
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.14
langsmith==0.4.56
lazy_loader==0.4
lxml==6.0.1
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
narwhals==2.1.2
nest-asyncio==1.6.0
networkx==3.4.2
newsapi-python==0.2.7
numpy==2.2.3
onnxruntime==1.21.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
openpyxl==3.1.5
orjson==3.11.5
ormsgpack==1.12.0
packaging==24.2
pandas==2.3.1
pillow==11.3.0
plotly==6.3.0
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.10.6
pydantic_core==2.27.2
pygame==2.6.1
pyglet==2.1.9
PyJWT==2.10.1
pymunk==6.9.0
pynput==1.8.1
pyparsing==3.2.3
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
python-dotenv==1.2.1
pytiled_parser==2.2.9
pytz==2025.2
PyYAML==6.0.3
reportlab==4.4.5
requests==2.32.3
requests-toolbelt==1.0.0
retrying==1.4.2
rsa==4.9.1
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.15.2
setuptools==80.9.0
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stix2==3.0.1
stix2-patterns==2.0.0
sympy==1.13.3
taxii2-client==2.3.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.2.18
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.2.0
urllib3==2.3.0
uuid_utils==0.12.0
websockets==15.0.1
Werkzeug==3.1.3
xxhash==3.6.0
zipp==3.23.0
zstandard==0.25.0
